Rename `transforms` directory to `common-tags`, add back empty `transforms` directory
