+++ packages/11ty/_plugins/common-tags/renderOneLine.js
